Backend/flask/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
pyjwt==1.7.1
Werkzeug==2.0.2
requests==2.21.0
psycopg2-binary
flask-cors
scikit-learn
numpy
